Mejoras globales

+++ AcademiaMusica/Pages/Alumnos/Index.cshtml.cs
@@ -16,13 +16,22 @@ namespace AcademiaMusica.Pages.Alumnos
 
         public IList<Alumno> Alumnos { get; set; }
 
-        public async Task OnGetAsync()
+        public async Task<IActionResult> OnGetAsync()
         {
+            var rol = HttpContext.Session.GetString("rol");
+            if (rol == "Alumno")
+                return RedirectToPage("/Index");
+
             Alumnos = await _db.GetAlumnos();
+            return Page();
         }
 
         public async Task<IActionResult> OnPostEliminarAsync(int id)
         {
+            var rol = HttpContext.Session.GetString("rol");
+            if (rol == "Alumno")
+                return RedirectToPage("/Index");
+
             await _db.DeleteAlumno(id);
             return RedirectToPage();
         }
