+++ AcademiaMusica/Pages/Instrumentos/Index.cshtml.cs
@@ -16,13 +16,22 @@ namespace AcademiaMusica.Pages.Instrumentos
 
         public IList<Instrumento> Instrumentos { get; set; }
 
-        public async Task OnGetAsync()
+        public async Task<IActionResult> OnGetAsync()
         {
+            var rol = HttpContext.Session.GetString("rol");
+            if (string.IsNullOrEmpty(rol))
+                return RedirectToPage("/Index");
+
             Instrumentos = await _db.GetInstrumentos();
+            return Page();
         }
 
         public async Task<IActionResult> OnPostEliminarAsync(int id)
         {
+            var rol = HttpContext.Session.GetString("rol");
+            if (string.IsNullOrEmpty(rol))
+                return RedirectToPage("/Index");
+
             await _db.DeleteInstrumento(id);
             return RedirectToPage();
         }
